no reason to go through an event here
i was only doing this before because libuv was using threads, and so i
needed a way to trigger redraws on the main thread. if there are other
reasons that we might want to delay redraws (to avoid blocking the event
loop, for instance), we should just do that by scheduling a 0-timeout
timer or something.